requirements.txt
streamlit==1.32.0
torch==2.0.1
fair-esm==1.0.3
numpy==1.24.3